+++ cut_cross_entropy/cce_lse_forward.py
@@ -50,21 +50,17 @@ def _cce_lse_forward_kernel(
     group_id = pid // num_pid_in_group
     first_pid_b = group_id * GROUP_B
     group_size_b = min(num_pid_b - first_pid_b, GROUP_B)
-    pid_b = first_pid_b + ((pid % num_pid_in_group) % group_size_b)
-    pid_v = (pid % num_pid_in_group) // group_size_b
+    pid_b = (first_pid_b + ((pid % num_pid_in_group) % group_size_b)).to(tl.int64)
+    pid_v = ((pid % num_pid_in_group) // group_size_b).to(tl.int64)
 
-    offs_b = pid_b * BLOCK_B + tl.arange(0, BLOCK_B)
+    offs_b = (pid_b * BLOCK_B + tl.arange(0, BLOCK_B)).to(tl.int64)
     if HAS_VALIDS:
-        offs_b = tl.load(Valids + stride_vb * offs_b.to(tl.int64), mask=offs_b < B, other=BMax)
+        offs_b = tl.load(Valids + stride_vb * offs_b, mask=offs_b < B, other=BMax).to(tl.int64)
 
-    offs_v = pid_v * BLOCK_V + tl.arange(0, BLOCK_V)
+    offs_v = (pid_v * BLOCK_V + tl.arange(0, BLOCK_V)).to(tl.int64)
     offs_d = tl.arange(0, BLOCK_D).to(tl.int64)
-    e_ptrs = E + (
-        offs_b[:, None].to(tl.int64) * stride_eb + offs_d[None, :].to(tl.int64) * stride_ed
-    )
-    c_ptrs = C + (
-        offs_v[None, :].to(tl.int64) * stride_cv + offs_d[:, None].to(tl.int64) * stride_cd
-    )
+    e_ptrs = E + (offs_b[:, None] * stride_eb + offs_d[None, :] * stride_ed)
+    c_ptrs = C + (offs_v[None, :] * stride_cv + offs_d[:, None] * stride_cd)
 
     accum = tl.zeros((BLOCK_B, BLOCK_V), dtype=tl.float32)
     for d in range(0, tl.cdiv(D, BLOCK_D)):
@@ -88,7 +84,7 @@ def _cce_lse_forward_kernel(
     tl.debug_barrier()
 
     if HAS_BIAS:
-        bias = tl.load(Bias + offs_v.to(tl.int64) * stride_biasv, mask=offs_v < V, other=0.0)
+        bias = tl.load(Bias + offs_v * stride_biasv, mask=offs_v < V, other=0.0)
         bias = bias.to(dtype=accum.dtype)
         accum += bias[None, :]
 
@@ -104,7 +100,7 @@ def _cce_lse_forward_kernel(
     e = tl.exp(logits - this_mx[:, None])
     this_lse = this_mx + tl.log(tl.sum(e, axis=1))
 
-    offs_b = pid_b * BLOCK_B + tl.arange(0, BLOCK_B)
+    offs_b = (pid_b * BLOCK_B + tl.arange(0, BLOCK_B)).to(tl.int64)
     o_mask = offs_b < B
 
     lse_ptrs = LSE + (stride_lse_b * offs_b)
